third_party/WebKit/LayoutTests/http/tests/push_messaging/pushevent-extendable-event.html

<!doctype html>
<html>
  <head>
    <title>Push API: PushEvent should be an ExtendableEvent.</title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src="../serviceworker/resources/test-helpers.js"></script>
  </head>
  <body>
    <script>
      // Tests that the PushEvent constructor is exposed in a ServiceWorker and
      // exposes the "data" and "waitUntil" members.
      service_worker_test(
          'resources/pushevent-extendable-event.js',
          'Exposure of the PushEvent object in a ServiceWorker.');
    </script>
  </body>
</html>
