third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/push_messaging/pushevent-extendable-event-expected.txt

This is a testharness.js-based test.
PASS Exposure of the PushEvent object in a ServiceWorker. 
PASS PushEvent is exposed and extends ExtendableEvent. 
Harness: the test ran to completion.
